# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "My API",
    "version": "1.0.0",
    "description": "API documentation for my Express application"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/auth/login": {
      "get": {
        "summary": "Render the login page",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Render the login form"
          }
        }
      },
      "post": {
        "summary": "Log in an existing user",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "The username of the user"
                  },
                  "password": {
                    "type": "string",
                    "description": "The password of the user"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "302": {
            "description": "Login successful, redirect to user role dashboard"
          },
          "401": {
            "description": "Invalid username or password"
          }
        }
      }
    },
...
    "/courses": {
      "get": {
        "summary": "Retrieve a list of all courses",
        "responses": {
          "200": {
            "description": "A list of courses"
          },
          "401": {
            "description": "Unauthorized access"
          }
        }
      }
    },
...
    "/teacher/courses": {
      "get": {
        "summary": "Retrieve all courses",
        "responses": {
          "200": {
            "description": "A list of courses"
          }
        }
      },
      "post": {
        "summary": "Create a new course",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Course created successfully"
          }
        }
      }
    },
    "/teacher/courses/{id}/edit": {
      "get": {
        "summary": "Render the form to edit a course",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the course to edit",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Render edit course form"
          }
        }
      },
      "post": {
        "summary": "Update a course",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the course to update",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Course updated successfully"
          },
          "404": {
            "description": "Course not found"
          }
        }
      }
    },
    "/teacher/courses/{id}/delete": {
      "post": {
        "summary": "Delete a course",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the course to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Course deleted successfully"
          },
          "404": {
            "description": "Course not found"
          }
        }
      }
    },
    "/teacher/quizzes": {
      "get": {
        "summary": "Retrieve all quizzes",
        "responses": {
          "200": {
            "description": "A list of quizzes"
          }
        }
      },
      "post": {
        "summary": "Create a new quiz",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "questions": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "question": {
                          "type": "string"
                        },
                        "options": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "correctAnswer": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Quiz created successfully"
          }
        }
      }
    },
...
    "/teacher/quizzes/{id}/edit": {
      "get": {
        "summary": "Render the form to edit a quiz",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the quiz to edit",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Render edit quiz form"
          }
        }
      }
    },
    "/teacher/quizzes/{id}": {
      "post": {
        "summary": "Update a quiz",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the quiz to update",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "questions": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "question": {
                          "type": "string"
                        },
                        "options": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "correctAnswer": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Quiz updated successfully"
          },
          "404": {
            "description": "Quiz not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a quiz",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the quiz to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Quiz deleted successfully"
          },
          "404": {
            "description": "Quiz not found"
          }
        }
      }
    },
    "/teacher/assignments/{id}": {
      "get": {
        "summary": "View details of a specific assignment",
        "tags": [
          "Teacher"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the assignment"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with assignment details"
          },
          "404": {
            "description": "Assignment not found"
          }
        }
      }
    },
    "/teacher/assignments": {
      "post": {
        "summary": "Create a new assignment",
        "tags": [
          "Teacher"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Assignment created successfully"
          },
          "400": {
            "description": "Bad request"
          }
        }
      },
      "get": {
        "summary": "View all assignments",
        "tags": [
          "Teacher"
        ],
        "responses": {
          "200": {
            "description": "Successful response with a list of assignments"
          }
        }
      }
    },
...
    "/teacher/gradeSubmission/{submissionId}": {
      "post": {
        "summary": "Grade a specific assignment submission",
        "tags": [
          "Teacher"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "submissionId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The submission ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "grade": {
                    "type": "integer",
                    "description": "Grade between 1 and 10"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Submission graded successfully"
          },
          "400": {
            "description": "Invalid grade or submission ID"
          }
        }
      }
    },
...
    "/teacher/gradeQuiz/{quizId}/{studentId}": {
      "post": {
        "summary": "Grade a specific quiz for a student",
        "tags": [
          "Teacher"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "quizId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The quiz ID"
          },
          {
            "in": "path",
            "name": "studentId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The student ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "grade": {
                    "type": "integer",
                    "description": "Grade for the quiz"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Quiz graded successfully"
          },
          "400": {
            "description": "Invalid quiz ID, student ID, or grade"
          }
        }
      }
    },
    "/teacher/assignments/{id}/grade": {
      "post": {
        "summary": "Grade a specific assignment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The assignment ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "grade": {
                    "type": "integer",
                    "description": "The grade to assign to the submission"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Assignment graded successfully"
          },
          "400": {
            "description": "Invalid assignment ID or grade"
          },
          "404": {
            "description": "Assignment not found"
          }
        },
        "tags": [
          "Teacher"
        ]
      }
    },
    "/teacher/upload": {
      "get": {
        "summary": "Render the file upload form",
        "responses": {
          "200": {
            "description": "Render upload form"
          }
        }
      },
      "post": {
        "summary": "Upload a file",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "File uploaded successfully"
          }
        }
      }
    },
...
    "/notifications/send": {
      "post": {
        "summary": "Send a notification to a student",
        "tags": [
          "Teacher"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "description": "The content of the notification"
                  },
                  "studentId": {
                    "type": "string",
                    "description": "The ID of the student"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Notification sent successfully"
          },
          "400": {
            "description": "Invalid input or missing parameters"
          }
        }
      }
    },
...
    "/teacher/assignments/{assignmentId}/submissions": {
      "get": {
        "summary": "View student submissions for a specific assignment",
        "tags": [
          "Teacher"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "assignmentId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the assignment"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with student submissions"
          },
          "404": {
            "description": "Assignment not found"
          }
        }
      }
    },
...
  },
...
}